requirements.txt
backports.functools-lru-cache==1.2.1
backports.ssl-match-hostname==3.5.0.1
cached-property==1.3.0
chardet==2.3.0
cycler==0.10.0
docker-compose==1.9.0
docker-py==1.10.6
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.0.4
ipaddress==1.0.16
IPy==0.81
jsonschema==2.5.1
matplotlib==1.5.3
numpy==1.11.3
pandas==0.19.2
Pillow==3.4.2
PyMySQL==0.7.9
pyparsing==2.1.10
PySocks==1.5.6
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.11
redis==2.10.5
repoze.lru==0.4
requests==2.10.0
scipy==0.18.1
six==1.10.0
texttable==0.8.1
urllib3==1.15.1
websocket-client==0.37.0
